# docs/swagger.yaml
components:
  schemas:
...
    Event:
      properties:
        application_end_time:
          format: date-time
          type: string
        application_start_time:
          format: date-time
          type: string
        event_end_time:
          format: date-time
          type: string
        event_start_time:
          format: date-time
          type: string
        game_version:
          $ref: '#/components/schemas/GameVersion'
        id:
          type: integer
        is_current:
          type: boolean
        is_locked:
          type: boolean
        is_main_event:
          type: boolean
        is_public:
          type: boolean
        max_size:
          type: integer
        name:
          type: string
        patch:
          type: string
        teams:
          items:
            $ref: '#/components/schemas/Team'
          type: array
          uniqueItems: false
        uses_medals:
          type: boolean
        waitlist_size:
          type: integer
      required:
      - application_end_time
      - application_start_time
      - event_end_time
      - event_start_time
      - game_version
      - id
      - is_current
      - is_locked
      - is_main_event
      - is_public
      - max_size
      - name
      - teams
      - uses_medals
      - waitlist_size
      type: object
...
    Team:
      properties:
        abbreviation:
          type: string
        allowed_classes:
          items:
            type: string
          type: array
          uniqueItems: false
        color:
          type: string
        discord_role_id:
          type: string
        event_id:
          type: integer
        id:
          type: integer
        name:
          type: string
      required:
      - allowed_classes
      - event_id
      - id
      - name
      type: object
    TeamCreate:
      properties:
        abbreviation:
          type: string
        allowed_classes:
          items:
            type: string
          type: array
          uniqueItems: false
        color:
          type: string
        discord_role_id:
          type: string
        id:
          type: integer
        name:
          type: string
      required:
      - allowed_classes
      - name
      type: object
...
    TeamSuggestion:
      properties:
        extra:
          type: string
        objective_id:
          type: integer
      type: object
    TeamUserCreate:
      properties:
        is_team_lead:
          type: boolean
        team_id:
          type: integer
        user_id:
          type: integer
      required:
      - user_id
      type: object
...
    User:
      properties:
        account_name:
          type: string
        discord_id:
          type: string
        discord_name:
          type: string
        display_name:
          type: string
        id:
          type: integer
        permissions:
          items:
            $ref: '#/components/schemas/Permission'
          type: array
          uniqueItems: false
        token_expiry_timestamp:
          format: date-time
          type: string
        twitch_id:
          type: string
        twitch_name:
          type: string
      required:
      - display_name
      - id
      - permissions
      type: object
...
    GameVersion:
      enum:
      - poe1
      - poe2
      type: string
      x-enum-varnames:
      - PoE1
      - PoE2
...
    Permission:
      enum:
      - admin
      - manager
      - objective_designer
      - submission_judge
      type: string
      x-enum-varnames:
      - PermissionAdmin
      - PermissionManager
      - PermissionObjectiveDesigner
      - PermissionSubmissionJudge
...
    SortedUser:
      properties:
        discord_id:
          type: string
        discord_name:
          type: string
        display_name:
          type: string
        is_team_lead:
          type: boolean
        poe_name:
          type: string
        team_id:
          type: integer
        user_id:
          type: integer
      required:
      - discord_id
      - discord_name
      - display_name
      - is_team_lead
      - poe_name
      - team_id
      - user_id
      type: object
  securitySchemes:
    BearerAuth:
      in: header
      name: Authorization
      type: apiKey
...
info:
  contact:
    email: Liberatorist@gmail.com
    name: Liberator
  description: This is the backend API for the BPL project.
  title: BPL Backend API
  version: "2.0"
openapi: 3.1.0
paths:
...
  /events/{event_id}:
    delete:
      description: Deletes an event
      operationId: DeleteEvent
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      tags:
      - event
    get:
      description: Fetches an event by id
      operationId: GetEvent
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
          description: OK
      security:
      - BearerAuth: []
      tags:
      - event
...
  /events/{event_id}/teams:
    get:
      description: Fetches all teams for an event
      operationId: GetTeams
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Team'
                type: array
          description: OK
      tags:
      - team
    put:
      description: Creates a team for an event
      operationId: CreateTeam
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/TeamCreate'
                description: Team to create
                summary: teamCreate
        description: Team to create
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
          description: Created
      security:
      - BearerAuth: []
      tags:
      - team
  /events/{event_id}/teams/{team_id}:
    delete:
      description: Deletes a team
      operationId: DeleteTeam
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      - description: Team Id
        in: path
        name: team_id
        required: true
        schema:
          type: integer
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      tags:
      - team
    get:
      description: Fetches a team by id
      operationId: GetTeam
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      - description: Team Id
        in: path
        name: team_id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
          description: OK
      tags:
      - team
...
  /events/{event_id}/teams/{team_id}/suggestions:
    get:
      description: Fetches all suggestions for your team for an event
      operationId: GetTeamSuggestions
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      - description: Team Id
        in: path
        name: team_id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TeamSuggestion'
                type: array
          description: OK
      security:
      - BearerAuth: []
      tags:
      - team
  /events/{event_id}/teams/{team_id}/suggestions/{objective_id}:
    delete:
      description: Deletes a suggestion for an objective for your team for an event
      operationId: DeleteObjectiveTeamSuggestion
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      - description: Team Id
        in: path
        name: team_id
        required: true
        schema:
          type: integer
      - description: Objective Id
        in: path
        name: objective_id
        required: true
        schema:
          type: integer
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      tags:
      - team
    put:
      description: Creates a suggestion for an objective for your team for an event
      operationId: CreateObjectiveTeamSuggestion
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      - description: Team Id
        in: path
        name: team_id
        required: true
        schema:
          type: integer
      - description: Objective Id
        in: path
        name: objective_id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/TeamSuggestion'
                description: Suggestion data
                summary: teamSuggestion
        description: Suggestion data
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                type: object
          description: Created
      security:
      - BearerAuth: []
      tags:
      - team
  /events/{event_id}/teams/users:
    get:
      description: Fetches all users of an event sorted by team and role
      operationId: GetSortedUsers
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/SortedUser'
                type: array
          description: OK
      tags:
      - team
      - user
    put:
      description: Adds users to teams
      operationId: AddUsersToTeams
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - items:
                  $ref: '#/components/schemas/TeamUserCreate'
                title: teamUsers
                type: array
        description: Users to add to teams
        required: true
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      tags:
      - team
      - user
...
  /users:
    get:
      description: Fetches all users
      operationId: GetAllUsers
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/User'
                type: array
          description: OK
      security:
      - BearerAuth: []
      tags:
      - user
  /users/{user_id}:
    get:
      description: Fetches a user by ID
      operationId: GetUserById
      parameters:
      - description: User Id
        in: path
        name: user_id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: OK
      tags:
      - user
    patch:
      description: Changes the permissions of a user
      operationId: ChangePermissions
      parameters:
      - description: User Id
        in: path
        name: user_id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - items:
                  $ref: '#/components/schemas/Permission'
                title: permissions
                type: array
        description: Permissions
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: OK
      security:
      - BearerAuth: []
      tags:
      - user
...